+++ crates/e2e-testing/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod asserts;
 pub mod controller;
 pub mod http_asserts;
-pub mod metadata_extractor;
+pub mod metadata;
 pub mod spin;
 pub mod spin_controller;
 pub mod testcase;
